app/requirements.txt
aiobotocore==1.3.0
aiohttp==3.6.2
aiopg==1.2.1
async-timeout==3.0.1
backoff==1.10.0
fastapi==0.65.1
opencv-contrib-python==4.5.2.52
pafy==0.5.5
Pillow==8.2.0
pytest==5.4.1
pytest-asyncio==0.10.0
python-dateutil==2.8.1
uvicorn==0.13.4
uvloop==0.15.2
youtube-dl==2020.3.24
asynctest==0.13.0
pytest-mock==2.0.0
pytest-aiohttp==0.3.0
pytest-cov==2.8.1
numpy==1.20.3
get-docker-secret==1.0.1
aiortsp==1.3.5
requests